% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bbd_logo.R
\name{add_bbd_logo}
\alias{add_bbd_logo}
\title{Bluebonnet Data \link{ggplot2} theme}
\usage{
add_bbd_logo(plot = ggplot2::last_plot(), position = "top right", scale = 1)
}
\arguments{
\item{plot}{The plot object passed to \code{add_bbd_logo}}

\item{position}{The position of the logo. Options are "top left", "top right", "bottom right", and "bottom left"}

\item{scale}{The scale of the BBD logo. Every single integer increase increases the scale multiplier (scale = 2 will double the size of the logo). Use decimals for fine-tuned scaling}
}
\description{
Add BBD logo to a single plot as grid object
Modified from https://gist.github.com/Ryo-N7/67ca1c364c342a82c4098918082ca445
}
