% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{bbd_colors}
\alias{bbd_colors}
\title{Function to extract `bbd_color_list` colors as hex codes}
\usage{
bbd_colors(...)
}
\arguments{
\item{...}{Character names of bbd_color_list colors}
}
\description{
Function to extract `bbd_color_list` colors as hex codes
}
