% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbd_fix_bars.R
\name{bbd_fix_bars}
\alias{bbd_fix_bars}
\title{Custom \link{ggplot2} theme for personal use by Bluebonnet Data}
\usage{
bbd_fix_bars(type = "continuous", axis = "y")
}
\arguments{
\item{type}{Data type. Options include "continuous" or "discrete"}

\item{axis}{Axis the data belong to. Options include "x" or "y"}
}
\description{
Drops tick axes and axes labels
}
