% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbd_font_install.R
\name{bbd_font_install}
\alias{bbd_font_install}
\title{Import and register Raleway}
\usage{
bbd_font_install(device = c("pdf", "win", "postscript"), force = FALSE)
}
\arguments{
\item{device}{Font device. Options are 'pdf', 'win', and 'postscript.' Defaults to 'pdf.' For more on your system's font device, see https://github.com/wch/extrafont}

\item{force}{Should the system force installation, even if Raleway is already present? You might select this option if you have since downloaded additional versions/faces of Raleway. Defaults to FALSE.}
}
\description{
\code{bbd_font_install} tests to see if default fonts imported and registered. If Raleway are not imported and registered, then \code{bbd_font_install} imports and registers Raleway with \href{https://github.com/wch/extrafont}{library(extrafont)}.
}
\details{
Note: Raleway must be installed on your computer for \code{bbd_font_install} to work.

Test to see if Raleway are imported and registered with \code{bbd_font_test()}.
}
