% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbd_font_test.R
\name{bbd_font_test}
\alias{bbd_font_test}
\title{Test for font import and registration}
\usage{
bbd_font_test()
}
\description{
\code{bbd_font_test} tests to see if default fonts are imported and registered.
}
\details{
Import and register fonts in R with bbd_font_install().
}
