% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{bbd_palettes}
\alias{bbd_palettes}
\alias{palette_bbd_main}
\alias{palette_bbd_diverging}
\alias{palette_bbd_sequential}
\title{List of CR Palettes}
\format{A \code{list}}
\usage{
bbd_palettes

palette_bbd_main

palette_bbd_diverging

palette_bbd_sequential
}
\description{
List of CR Palettes

Bluebonnet Data custom color palettes
}
\concept{CR palettes}
\keyword{datasets}
