% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bbd_theme.R
\name{set_bbd_theme}
\alias{set_bbd_theme}
\title{The BBD \link{ggplot2} theme}
\usage{
set_bbd_theme(style = "print")
}
\arguments{
\item{style}{The default theme style for the R session. Options are "print" or "Texas".}
}
\description{
\code{set_bbd_theme} provides a \link{ggplot2} theme formatted according to the
BBD website.
}
