% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bbd_print.R
\name{theme_bbd_print}
\alias{theme_bbd_print}
\title{A print \link{ggplot2} theme formatted in BBD style}
\usage{
theme_bbd_print(
  base_size = 11,
  base_family = "raleway",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base line}

\item{base_rect_size}{rectangle sizes}
}
\description{
\code{theme_bbd} provides a \link{ggplot2} theme formatted according to BBD
website
}
