% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bbd_texas.R
\name{theme_bbd_texas}
\alias{theme_bbd_texas}
\title{A Texas shaped \link{ggplot2} theme formatted in BBD style}
\usage{
theme_bbd_texas(
  scale = "continuous",
  base_size = 12,
  base_family = "raleway",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{scale}{scale}

\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base line}

\item{base_rect_size}{base rectangle sizes}
}
\description{
\code{theme_bbd} provides a \link{ggplot2} theme formatted according to the BBD
website
}
