% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undo_bbd_theme.R
\name{undo_bbd_theme}
\alias{undo_bbd_theme}
\title{Custom \link{ggplot2} theme for personal use by Bluebonnet Data}
\usage{
undo_bbd_theme()
}
\description{
Resets the \link{ggplot2} theme to its default settings
}
