% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_palette.R
\name{view_palette}
\alias{view_palette}
\title{Custom \link{ggplot2} theme for personal use by Bluebonnet Data}
\usage{
view_palette(palette = palette_bbd_main)
}
\arguments{
\item{palette}{A \code{palette_bbd_*} vector from \code{library(bbdata)}.
Options are \code{palette_bbd_main}, \code{palette_bbd_diverging}, \code{palette_bbd_politics}, and \code{palette_bbd_sequential}}
}
\description{
\code{view_palette} displays the colors and hexadecimal codes for \code{palette_bbd_*} vectors.
These palettes were generated on www.coolors.co/.
}
