% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbd_font_install.R
\name{bbd_font_install}
\alias{bbd_font_install}
\title{Import and register Raleway & IBM Plex Sans font}
\usage{
bbd_font_install()
}
\description{
\code{bbd_font_install} tests to see if default fonts imported and registered. If Raleway & IBM Plex Sans are not imported and registered, then \code{bbd_font_install} imports and registers Raleway & IBM Plex Sans with \href{https://github.com/wch/extrafont}{library(extrafont)}.
}
\details{
Note: Raleway and IBM Plex Sans must be installed on your computer for \code{bbd_font_install} to work.

Test to see if Raleway and IBM Plex Sans are imported and registered with \code{bbd_font_test()}.
}
