% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bbd_theme.R
\name{set_bbd_theme}
\alias{set_bbd_theme}
\title{The BBD \link{ggplot2} theme}
\usage{
set_bbd_theme(style = "print", font = "raleway")
}
\arguments{
\item{style}{The default theme style for the R session. Options are "print" or "map".}

\item{font}{The font for plot labels, axes, and titles. Options are "raleway" (Raleway) and "ibm" (IBM Plex Sans).}
}
\description{
\code{set_bbd_theme} provides a \link{ggplot2} theme formatted according to the
BBD website.
}
